1. Verify Authorisation Added 2. User Controller Second Level Authorization Added

+++ app/controllers/users_controller.rb
@@ -29,34 +29,54 @@ class UsersController < ApplicationController
 
 		user = User.find(params[:id])
 
-		if user.update(user_params)
+		if current_user.id == user.id || current_user.auth_level == "superadmin"
 
-			flash[:notice] = "Successfully Updated #{user.first_name}!"
-			redirect_to action: "show", id: params[:id]
+			if user.update(user_params)
+
+				flash[:notice] = "Successfully Updated #{user.first_name}!"
+				redirect_to action: "show", id: params[:id]
+
+			else
+
+				flash[:notice] = 'Unsuccessful! Invalid Change!'
+				redirect_to action: "show", id: params[:id]
+
+			end
 
 		else
 
-			flash[:notice] = 'Unsuccessful! Invalid Change!'
-			redirect_to action: "show", id: params[:id]
+			flash[:notice] = "You are not Authorized to do this!"
+			redirect_back(fallback_location: root_path)
 
 		end
+
 	end
 
 	def destroy
 
 		user = User.find(params[:id])
 
-		if user.destroy
+		if current_user.id == user.id || current_user.auth_level == "superadmin"
 
-			flash[:notice] = "Thank you! We're sorry you had to go!"
-			redirect_to '/'
+			if user.destroy
+
+				flash[:notice] = "Thank you! We're sorry you had to go!"
+				redirect_to '/'
+
+			else
+
+				flash[:notice] = "Error! Unable to delete account. Try again later."
+				redirect_to action: "show", id: params[:id]
+
+			end
 
 		else
 
-			flash[:notice] = "Error! Unable to delete account. Try again later."
-			redirect_to action: "show", id: params[:id]
+			flash[:notice] = "You are not Authorized to do this!"
+			redirect_back(fallback_location: root_path)
 
 		end
+
 	end
 
 	private
